extract idempotency lookup route
